scx_rusty: factor out per-cpu variable logic
